requirements.txt
streamlit>=1.30
openai>=1.0.0
python-dotenv